requirements.txt
boto3==1.34.23
botocore==1.34.23
jmespath==1.0.1
psycopg2==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
s3transfer==0.10.0
six==1.16.0
urllib3==2.0.7
